move _largest_component_mask to _utils.py
