+++ src/cedalion/geometry/photogrammetry/anonymization/preprocessing.py
@@ -14,7 +14,6 @@ landmark detection and mask building:
 import logging
 
 import numpy as np
-import trimesh
 
 import cedalion.dataclasses as cdc
 import cedalion.typing as cdt
@@ -22,6 +21,7 @@ import cedalion.typing as cdt
 from ._utils import (
     _apply_affine,
     _ear_midpoint,
+    _largest_component_mask,
     _rebuild_mesh,
     _reindex_faces,
     _upper_head_centroid,
@@ -80,52 +80,6 @@ def orient_y_anterior(
     )
 
     return rotated_surface, rotated_nasion, R
-
-
-def _largest_component_mask(mesh: trimesh.Trimesh) -> np.ndarray:
-    """Boolean vertex mask selecting the largest connected component.
-
-    Einstar scans contain floating fragments (loose triangles, cable shreds,
-    background patches) that drag vertex extrema off into empty space; strip
-    them first.
-
-    Uses ``trimesh.graph.connected_component_labels`` on face adjacency rather
-    than ``mesh.split``: split allocates a ``Trimesh`` per component and can
-    OOM on scans with thousands of fragments.
-
-    Einstar OBJs duplicate vertices along UV seams; ``Trimesh.merge_vertices``
-    preserves seams (to keep textures), so face adjacency on the raw mesh
-    over-fragments the head. ``trimesh.grouping.unique_rows`` does
-    position-only merging for connectivity analysis.
-
-    Args:
-        mesh: A ``trimesh.Trimesh`` instance.
-
-    Returns:
-        Boolean array of shape ``(n_vertices,)``. All-True if the mesh is
-        empty or already a single connected component.
-    """
-    n_verts = len(mesh.vertices)
-    n_faces = len(mesh.faces)
-    if n_faces == 0:
-        return np.ones(n_verts, dtype=bool)
-
-    _, inverse = trimesh.grouping.unique_rows(np.asarray(mesh.vertices))
-    canonical_faces = inverse[mesh.faces]
-    adjacency = trimesh.graph.face_adjacency(faces=canonical_faces)
-    face_labels = trimesh.graph.connected_component_labels(
-        adjacency, node_count=n_faces
-    )
-    counts = np.bincount(face_labels)
-    if len(counts) <= 1:
-        return np.ones(n_verts, dtype=bool)
-    biggest_label = int(np.argmax(counts))
-    face_mask = face_labels == biggest_label
-
-    kept_vidx = np.unique(mesh.faces[face_mask])
-    mask = np.zeros(n_verts, dtype=bool)
-    mask[kept_vidx] = True
-    return mask
 
 
 @cdc.validate_schemas
